Rename Pipeline -> Workflow, rename top-level 'actions' to 'workflow'
